implement web crawler in main function
hardcoded but it works. Anyways, the provided command-line argument is parsed and the domain is extracted. Added counter variables to track number of successful and failed URL accesses. The web cralwer is invoked using the extracted domain, which results in it iterating through URL states. Obviously, added handling for the URL state to increment counters. Finally, a progress display at the very end